requirements.txt
fastapi
uvicorn
openai
sentence-transformers
faiss-cpu
PyPDF2
python-multipart
